server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis[hiredis]==5.0.8
pydantic==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psutil==6.0.0
aiofiles==24.1.0
python-multipart==0.0.9
httpx==0.27.2
